+++ R/nchar.R
@@ -32,7 +32,7 @@
 #'
 #' @inheritParams strip_ctl
 #' @inheritParams base::nchar
-#' @note the `keepNA` parameter is ignored for R < 3.2.
+#' @note the `keepNA` parameter is ignored for R < 3.2.2.
 #' @export
 #' @param type character string, one of "chars", or "width".  For byte counts
 #'   use [base::nchar].
@@ -91,8 +91,8 @@ nchar_ctl <- function(
   type <- valid.types[type.int]
   stripped <- strip_ctl(x, strip=strip, warn=warn)
 
-  R.ver.gte.3.2 <- R.ver.gte.3.2 # "import" symbol from namespace
-  if(R.ver.gte.3.2) nchar(stripped, type=type, allowNA=allowNA, keepNA=keepNA)
+  R.ver.gte.3.2.2 <- R.ver.gte.3.2.2 # "import" symbol from namespace
+  if(R.ver.gte.3.2.2) nchar(stripped, type=type, allowNA=allowNA, keepNA=keepNA)
   else nchar(stripped, type=type, allowNA=allowNA) # nocov
 }
 #' @export