+++ tests/unitizer/tabs.R
@@ -6,6 +6,7 @@ unitizer_sect('simple tabs', {
   tabs_as_spaces(string)
   tabs_as_spaces(string, c(2, 3, 4, 5, 8))
   tabs_as_spaces(string, c(2, 8))
+  tabs_as_spaces(1:3)
 })
 unitizer_sect('newlines', {
   string.n <- paste0(
@@ -21,4 +22,8 @@ unitizer_sect('corner cases', {
   tabs_as_spaces('\t')
   tabs_as_spaces('\n')
   tabs_as_spaces(c(string, string, string))
+})
+unitizer_sect('bad inputs', {
+  tabs_as_spaces(string, warn=1:3)
+  tabs_as_spaces(string, tab.stops='hello')
 })