+++ NEWS.md
@@ -1,5 +1,23 @@
 # fansi Release Notes
 
+## v0.3.0
+
+* `fansi::set_knit_hooks` makes it easy to automatically convert ANSI CSI SGR
+  sequences to HTML in Rmarkdown documents.  We also add a vignette that
+  demonstrates how to do this.
+* [#53](https://github.com/brodieG/fansi/issues/53): fix for systems where
+  'char' is signed (found and fixed by @QuLogic).
+* [#52](https://github.com/brodieG/fansi/issues/52): fix bad compilation under
+  ICC (@kazumits).
+* [#51](https://github.com/brodieG/fansi/issues/51): documentation improvements
+  (@krlmlr).
+* [#50](https://github.com/brodieG/fansi/issues/50): run tests on R 3.1 - 3.4
+  tests for the rc branch only (@krlmlr).
+* [#48](https://github.com/brodieG/fansi/issues/48): malformed call to error
+  in FANSI_check_enc (@msannell).
+* [#47](https://github.com/brodieG/fansi/issues/47): compatibility with R
+  versions 3.2.0 and 3.2.1 (@andreadega).
+
 ## v0.2.3
 
 * [#45](https://github.com/brodieG/fansi/issues/45): add capability to run under